Implement traits for individual configuration of each pass
The LSP riscv_get_diagnostics function currently always uses
the default configuration. This should be made configurable in the
future. Also, only the DotCFGGenerationPass is configurable via
the CLI currently (it is opt-in based on whether the --dot-cfg
option is passed). Other passes have PassConfiguration
implementations, but these are not integrated with the CLI yet.