Add --dot-cfg-interprocedural boolean flag
If this flag is passed as an argument to the lint command in
addition to the --dot-cfg=<PATH> flag, the generated CFG
will be interprocedural. The difference from the default
(intraprocedural) CFG is that, in the interprocedural CFG,
function calls terminate basic blocks and dashed edges are drawn
for function calls/returns.